requirements.txt
# ========================================
# DJANGO & WEB FRAMEWORK
# ========================================
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# ========================================
# BACKEND - VerifiK
# ========================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Redis & Cache
redis==5.2.0
celery==5.4.0

# AI/Computer Vision
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.0  # YOLOv8
opencv-python==4.10.0.84
numpy==2.1.3
pillow==11.0.0

# API & Integration
httpx==0.27.2
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
playwright==1.48.0

